requirements.txt
opencv-python==4.6.0.66
mediapipe>=0.10.5
numpy==1.23.4
matplotlib==3.6.2
pygaze==1.2.2
pandas==1.5.1
scikit-learn==1.1.3
pillow==9.3.0
imutils==0.5.4 